requirements.txt
asttokens==2.0.5
backcall==0.2.0
decorator==5.1.1
et-xmlfile==1.1.0
executing==0.8.3
ipdb==0.13.9
ipython==8.1.1
jedi==0.18.1
matplotlib-inline==0.1.3
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
PyQRCode==1.2.1
python-dateutil==2.8.2
pytz==2021.3
qrcode==7.3.1
six==1.16.0
stack-data==0.2.0
toml==0.10.2
traitlets==5.1.1
wcwidth==0.2.5
